remove internal.ts
split up code into IScale.ts, ICommonHeatMapOptions.ts, defaultUtils.ts
