+++ src/heatmap/ICommonHeatMapOptions.ts
@@ -1,0 +1,27 @@
+import {IVisInstanceOptions} from 'phovea_core/src/vis';
+
+export interface ICommonHeatMapOptions extends IVisInstanceOptions {
+  /**
+   * @default 10
+   */
+  initialScale?: number;
+  /**
+   * @default derived from value
+   */
+  color?: string[];
+  /**
+   * @default derived from value
+   */
+  domain?: (number|string)[];
+
+  /**
+   * missing value color
+   * @default magenta
+   */
+  missingColor?: string;
+
+  /**
+   * defines the rendering mode, e.g. influencing how the selection is drawn
+   */
+  mode?: 'lg'|'sm';
+}
